backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
python-docx
pandas
numpy
google-generativeai
groq
openai
python-pptx
python-dotenv